pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "mmsg"
authors= [{name = "Franz Louis Cesista"}]
description = "Multimodal Structured Generation"
requires-python = ">=3.8"
keywords=[
    "machine learning",
    "deep learning",
    "language models",
    "computer vision",
    "structured generation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "interegular",
    "jsonschema",
    "outlines @ git+https://github.com/outlines-dev/outlines@main",
    "Pillow",
    "referencing",
    "requests",
    "torch",
]
dynamic = ["version"]

[project.optional-dependencies]
test = [
    "modal",
    "numpy",
    "term-image",
    "pre-commit",
    "transformers @ file:///transformers",
    # "transformers",
]

[project.urls]
homepage = "https://github.com/leloykun/mmsg"
repository = "https://github.com/leloykun/mmsg"

[project.readme]
file="README.md"
content-type = "text/markdown"

[tool.setuptools]
packages = ["mmsg"]

[tool.setuptools.package-data]
"mmsg" = ["py.typed"]

[tool.setuptools_scm]
write_to = "mmsg/_version.py"

[tool.mypy]
exclude=[]
enable_incomplete_feature = ["Unpack"]

[tool.flake8]
max-line-length = 89

[tool.black]
line-length = 89

[[tool.mypy.overrides]]
module = [
    "interegular.*",
    "jsonschema.*",
    "numpy.*",
    "outlines.*",
    "PIL.*",
    "referencing.*",
    "requests.*",
    "torch.*",
    "transformers.*",
]
ignore_missing_imports = true
